Fix bug where a user could keep submitting name changes and it would rollback to one of their previous pending changes. Add redirect and flash message to approve and reject actions.

+++ app/controllers/approvals_controller.rb
@@ -5,16 +5,28 @@ class ApprovalsController < ApplicationController
   end
   
   def approve
-    version = PaperTrail::Version.find(params[:id])
-    model = version.item_type.constantize
-    model.paper_trail_off!
-    version.reify.save
-    model.paper_trail_on!
-    version.update_attributes(status: "approved")
+    version = PaperTrail::Version.find_by_id(params[:id])
+    if version
+      model = version.item_type.constantize
+      model.paper_trail_off!
+      version.reify.save
+      model.paper_trail_on!
+      version.update_attributes(status: "approved")
+      flash[:success] = "Version #{params[:id]} has been approved"
+    else
+      flash[:error] = "Version #{params[:id]} does not exist or has already been rejected"
+    end
+    redirect_to approvals_url
   end
   
   def reject
     version = PaperTrail::Version.find(params[:id])
-    version.destroy
+    if (version.status == "approved")
+      flash[:error] = "Version #{params[:id]} has already been approved"
+    else
+      version.destroy
+      flash[:info] = "Version #{params[:id]} has been rejected"
+    end
+    redirect_to approvals_url
   end
 end
